% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopkinsdown.R
\docType{package}
\name{hopkinsdown}
\alias{hopkinsdown}
\alias{hopkinsdown-package}
\title{unimelb: A package for creating University of Melbourne theses using R
Markdown}
\description{
unimelb: A package for creating University of Melbourne theses using R
Markdown
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

